app/requirements.txt
aiogram==3.10.0
aiohttp==3.9.4
beautifulsoup4==4.12.3
tortoise-orm==0.20.0
asyncpg==0.29.0  # Для PostgreSQL
python-dotenv==1.0.0
apscheduler==3.10.4  # Для планирования напоминаний
gigachat